Add a generic call-signature type, via a type-level uncurry function!
Getting this to type check was pretty rough, but I'm quite happy with
the eventual solution.  Also, I'm using type-level strings to get
unique signature types and recover a value-level string, without
having to duplicate the string at the value level.  This is a minor
savings in practice, but it's cool that it works (via Iavor's TypeLits
extension).