Add most of a simple logger.
From the comments:

  ----------------------------------------------------------------
  -- An event logger with a simple interface.
  --
  -- Specify the function you want to trace, along with a 'Symbol' to
  -- uniquely tag it (not implemented here), and get back a tracer for
  -- your function.  The pattern for lightweight logging is then
  --
  --   f :: t
  --   f = e
  --
  -- becomes
  --
  --   f , f' :: t
  --   f = simpleLogger "f" f'
  --   f' = e
  --
  -- creating a knot, where recursive calls in 'e' are still to 'f' and
  -- hence trigger recursive logging.

The handling of 'Symbols' and related creation of 'SimpleCall'
signatures is not implemented yet, but that part should be relatively
simple. Here we have the implementation of a "collecting" call that
wraps a function so that it's accumulated (uncurried) parameters get
returned with it's result.  E.g. a function of type

  a1 -> ... -> an -> m r

turns into a function of type

  a1 -> ... -> an -> ((a1 , (...(an , ()))) , m r)

internally, and then the accumulation of arguments

  (a1 , (...(an , ())))

can be used to construct a 'BeginCall' log event.

The current version combines the accumulation with the logging, but it
might be nicer to abstract out the pattern into a separate collection
function that's used by a simpler logging function.  However, that
might complicate the types even more ... we'll see.