CP: Implement 'simpleLogger', up to realizing my use of 'Typeable' is doomed.
I've use a 'Typeable' constraint to make the parser type safe, by
allowing it to 'cast' the 'Signature' call tag in order to be sure
that a 'BeginCall' and an 'EndCall' really correspond to the same call
tag.  However, there's no hope of making the

  Proxy (SimpleCall _ _ _ _)

an instance of 'Typeable', since that that makes the library hard to
use, since the user then has a 'Typeable' constraint on their function
signatures.  It turns out that the use of non-'*'-kinded type params
also precludes deriving 'Typeable', and so the use of 'Symbol' is also
an issue here, but that's not the real problem.

However, it's not hard to work around this: just wait until after the
function returns to log it's 'EndCall', and put all the arguments in
the end call.  I'll lose the identification of ill-formed log-event
streams, which was based on imbalanced call tags -- using 'cast' to
check for type equality, which doesn't work now that we lose
'Typeable' on the call tags -- but this doesn't seem like a big loss,
especially since I expect to only use auto generated event streams,
and so they'll only be messed up if my auto logging code is messed up.

It occurs to me now that I shouldn't in general assume we're in a
writer monad. I was thinking that's the only way someone would log,
but of course mutable logging in 'State' or 'IO' is also reasonable.
Since 'IO' is not defined as a transform of 'Identity', I 'IO'
instances for 'UncurryM' and friends. To come ...