Add constraint implication and 'LogTree' coercion.
This may be the trickiest code so far!  It was conceptually hard to
come up with an implementable definition of constraint implication (I
ended up with 2 equivalent defs!).  Once we have the definitions, it's
easy to see how they are used to write coercions, but in practice it's
hard to make the coercions type check, because they involve a lot of
careful manipulation of types, which are mostly implicit in Haskell.
So, I needed a lot of type signatures and scoped type variables, and
careful supplying of arguments to trigger type quantifier manipulation
(or at least, that's what I think is going on; e.g. writing
'implies2 impl = f impl' instead of 'implies2 = f').

Now that we have 'LogTree' coercion, it should be possible to log once
and then post process via many processors.  So, I better start
implementing more processors and logging examples ...