CP: Add the 'HetCall' call and the translation function 'heterogenize'.
The 'HetCall c tag' is a 'Signature' for which all parts are wrapped
in 'H c'; the idea is to facilitate generic processor instances which
use the constraint 'c' to pretty print the contents, but which don't
care which particular function is being processed.  The simplest case
is probably to make 'c ~ Show'.

The 'tag' is there so that we can write more than one processor
instance for the same class 'c'.  E.g., the library could provide a

  UnixTree (HetCall Show "default")

instance and a user could write there own

  UnixTree (HetCall Show "my fancy instance")

if they didn't like the default.