Add 'HFold c [H c]' instance and 'unH' for processing 'HetCall's.
In a 'LogTree (HetCall c t)' node all the parts are wrapped in 'H c's,
with the args in a '[H c]'.  The 'unH' function abstracts the only
reasonable thing you can do with an 'H c' value: unwrap it and apply
some function that uses the constraint 'c'.  The 'HFold c [H c]'
instance is essentially a regular fold where the folding function is
appropriately composed with 'unH' to go under the 'H c's.