Add 'unsafeTrace', a version of 'trace' for pure functions.
The pure function 'f' in 'unsafeTrace f' still needs to be open
recursive or mutual recursive, so that we can get inside the knot, but
by using 'unsafePerformIO' we can lift the need for a monad.

I don't completely understand why I need to force the return value in
'trace'; I don't need to in the "safe" version.