CP: use default 'UnixTree' backend in the type-inference experiment.
The current code breaks the "self contained" test, and using the
default 'Show'-based backends is pretty ugly here.  Realistically,
we'd want a more generic default backend which did not default to
'Show'. Then, if we add a "mode" to the 'Pretty' class, we can have
latex and plain text modes, and then use 'Pretty' as the class used
with the default backend.

With 'Show', the printout is unwieldy; see
experiments/proof-trees/out.txt.